pyincept-37: Consolidate the classes FileRenderer and FileRendererBase under FileRenderer.
